Add a TLS server and client
Add a TLS server and client that set up a mutually authenticated TLS
connection.

The server is useful for providing a TLS server that will authenticate a
client, giving a client something to authenticate to (as most TLS
servers as deployed on the net don't authenticate clients).

The client is useful for testing that the server works, independently of
any other client code. It also demonstrates how to do client
authentication from Go.

Keys and certificates are added for both client and server.